telegram_bot/requirements.txt
ccxt>=4.0.0
aiogram>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
aiofiles>=23.0.0
matplotlib>=3.5.0